Add base SimpleBook installation
Simplebook has two services which need to run in order for the plugin to
work. The first is an API server which will accept new book requests,
the second is a node worker which actually generates PDFs.

This makes it possible to install both services, but does not add the
SimpleBook plugin to any competitions.

Issue #10